Bulk clean-up (mostly deletion) of Jitter Buffer implementations
