+++ android_streaming_client/doc/com/c77/androidstreamingclient/lib/rtp/NoDelayRtpMediaBuffer.html
@@ -86,7 +86,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.c77.androidstreamingclient.lib.rtp.NoDelayRtpMediaBuffer</li>
+<li>com.c77.androidstreamingclient.lib.rtp.buffer.MinDelayRtpMediaBuffer</li>
 </ul>
 </li>
 </ul>
@@ -205,7 +205,7 @@ implements com.biasedbit.efflux.session.RtpSessionDataListener, <a href="../../.
 <li class="blockList">
 <h4>CONFIG_TIMEOUT_MS</h4>
 <pre>public static final&nbsp;java.lang.String CONFIG_TIMEOUT_MS</pre>
-<dl><dt><span class="strong">See Also:</span></dt><dd><a href="../../../../../constant-values.html#com.c77.androidstreamingclient.lib.rtp.NoDelayRtpMediaBuffer.CONFIG_TIMEOUT_MS">Constant Field Values</a></dd></dl>
+<dl><dt><span class="strong">See Also:</span></dt><dd><a href="../../../../../constant-values.html#com.c77.androidstreamingclient.lib.rtp.buffer.NoDelayRtpMediaBuffer.CONFIG_TIMEOUT_MS">Constant Field Values</a></dd></dl>
 </li>
 </ul>
 </li>
