+++ android_streaming_client/doc/com/c77/androidstreamingclient/lib/rtp/OriginalRtpMediaExtractor.html
@@ -86,7 +86,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.c77.androidstreamingclient.lib.rtp.OriginalRtpMediaExtractor</li>
+<li>com.c77.androidstreamingclient.lib.rtp.RtpMediaExtractor</li>
 </ul>
 </li>
 </ul>
