+++ android_streaming_client/doc/com/c77/androidstreamingclient/lib/rtp/RtpMediaBuffer.State.html
@@ -89,7 +89,7 @@
 <li>java.lang.Enum&lt;<a href="../../../../../com/c77/androidstreamingclient/lib/rtp/RtpMediaBuffer.State.html" title="enum in com.c77.androidstreamingclient.lib.rtp">RtpMediaBuffer.State</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.c77.androidstreamingclient.lib.rtp.RtpMediaBuffer.State</li>
+<li>com.c77.androidstreamingclient.lib.rtp.buffer.RtpMediaBuffer.State</li>
 </ul>
 </li>
 </ul>
