+++ android_streaming_client/doc/com/c77/androidstreamingclient/lib/rtp/RtpMediaBufferWithJitterAvoidance.html
@@ -120,7 +120,7 @@ implements <a href="../../../../../com/c77/androidstreamingclient/lib/rtp/RtpMed
 </a>
 <h3>Nested Class Summary</h3>
 <ul class="blockList">
-<li class="blockList"><a name="nested_classes_inherited_from_class_com.c77.androidstreamingclient.lib.rtp.RtpMediaBuffer">
+<li class="blockList"><a name="nested_classes_inherited_from_class_com.c77.androidstreamingclient.lib.rtp.buffer.RtpMediaBuffer">
 <!--   -->
 </a>
 <h3>Nested classes/interfaces inherited from interface&nbsp;com.c77.androidstreamingclient.lib.rtp.<a href="../../../../../com/c77/androidstreamingclient/lib/rtp/RtpMediaBuffer.html" title="interface in com.c77.androidstreamingclient.lib.rtp">RtpMediaBuffer</a></h3>
