+++ android_streaming_client/doc/com/c77/androidstreamingclient/lib/rtp/RtpMediaJitterBuffer.html
@@ -86,7 +86,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.c77.androidstreamingclient.lib.rtp.RtpMediaJitterBuffer</li>
+<li>com.c77.androidstreamingclient.lib.rtp.buffer.TimeWindowRtpMediaBuffer</li>
 </ul>
 </li>
 </ul>
@@ -120,7 +120,7 @@ implements <a href="../../../../../com/c77/androidstreamingclient/lib/rtp/RtpMed
 </a>
 <h3>Nested Class Summary</h3>
 <ul class="blockList">
-<li class="blockList"><a name="nested_classes_inherited_from_class_com.c77.androidstreamingclient.lib.rtp.RtpMediaBuffer">
+<li class="blockList"><a name="nested_classes_inherited_from_class_com.c77.androidstreamingclient.lib.rtp.buffer.RtpMediaBuffer">
 <!--   -->
 </a>
 <h3>Nested classes/interfaces inherited from interface&nbsp;com.c77.androidstreamingclient.lib.rtp.<a href="../../../../../com/c77/androidstreamingclient/lib/rtp/RtpMediaBuffer.html" title="interface in com.c77.androidstreamingclient.lib.rtp">RtpMediaBuffer</a></h3>
@@ -232,7 +232,7 @@ implements <a href="../../../../../com/c77/androidstreamingclient/lib/rtp/RtpMed
 <li class="blockList">
 <h4>DEBUGGING_PROPERTY</h4>
 <pre>public static final&nbsp;java.lang.String DEBUGGING_PROPERTY</pre>
-<dl><dt><span class="strong">See Also:</span></dt><dd><a href="../../../../../constant-values.html#com.c77.androidstreamingclient.lib.rtp.RtpMediaJitterBuffer.DEBUGGING_PROPERTY">Constant Field Values</a></dd></dl>
+<dl><dt><span class="strong">See Also:</span></dt><dd><a href="../../../../../constant-values.html#com.c77.androidstreamingclient.lib.rtp.buffer.RtpMediaJitterBuffer.DEBUGGING_PROPERTY">Constant Field Values</a></dd></dl>
 </li>
 </ul>
 <a name="FRAMES_WINDOW_PROPERTY">
@@ -242,7 +242,7 @@ implements <a href="../../../../../com/c77/androidstreamingclient/lib/rtp/RtpMed
 <li class="blockList">
 <h4>FRAMES_WINDOW_PROPERTY</h4>
 <pre>public static final&nbsp;java.lang.String FRAMES_WINDOW_PROPERTY</pre>
-<dl><dt><span class="strong">See Also:</span></dt><dd><a href="../../../../../constant-values.html#com.c77.androidstreamingclient.lib.rtp.RtpMediaJitterBuffer.FRAMES_WINDOW_PROPERTY">Constant Field Values</a></dd></dl>
+<dl><dt><span class="strong">See Also:</span></dt><dd><a href="../../../../../constant-values.html#com.c77.androidstreamingclient.lib.rtp.buffer.RtpMediaJitterBuffer.FRAMES_WINDOW_PROPERTY">Constant Field Values</a></dd></dl>
 </li>
 </ul>
 </li>
