+++ android_streaming_client/doc/constant-values.html
@@ -83,7 +83,7 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.NoDelayRtpMediaBuffer.CONFIG_TIMEOUT_MS">
+<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.buffer.MinDelayRtpMediaBuffer.CONFIG_TIMEOUT_MS">
 <!--   -->
 </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
 <td><code><a href="com/c77/androidstreamingclient/lib/rtp/NoDelayRtpMediaBuffer.html#CONFIG_TIMEOUT_MS">CONFIG_TIMEOUT_MS</a></code></td>
@@ -102,21 +102,21 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.OriginalRtpMediaExtractor.CSD_0">
+<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.RtpMediaExtractor.CSD_0">
 <!--   -->
 </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
 <td><code><a href="com/c77/androidstreamingclient/lib/rtp/OriginalRtpMediaExtractor.html#CSD_0">CSD_0</a></code></td>
 <td class="colLast"><code>"csd-0"</code></td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.OriginalRtpMediaExtractor.CSD_1">
+<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.RtpMediaExtractor.CSD_1">
 <!--   -->
 </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
 <td><code><a href="com/c77/androidstreamingclient/lib/rtp/OriginalRtpMediaExtractor.html#CSD_1">CSD_1</a></code></td>
 <td class="colLast"><code>"csd-1"</code></td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.OriginalRtpMediaExtractor.DURATION_US">
+<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.RtpMediaExtractor.DURATION_US">
 <!--   -->
 </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
 <td><code><a href="com/c77/androidstreamingclient/lib/rtp/OriginalRtpMediaExtractor.html#DURATION_US">DURATION_US</a></code></td>
@@ -234,14 +234,14 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.RtpMediaJitterBuffer.DEBUGGING_PROPERTY">
+<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.buffer.TimeWindowRtpMediaBuffer.DEBUGGING_PROPERTY">
 <!--   -->
 </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
 <td><code><a href="com/c77/androidstreamingclient/lib/rtp/RtpMediaJitterBuffer.html#DEBUGGING_PROPERTY">DEBUGGING_PROPERTY</a></code></td>
 <td class="colLast"><code>"DEBUGGING"</code></td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.RtpMediaJitterBuffer.FRAMES_WINDOW_PROPERTY">
+<td class="colFirst"><a name="com.c77.androidstreamingclient.lib.rtp.buffer.TimeWindowRtpMediaBuffer.FRAMES_WINDOW_PROPERTY">
 <!--   -->
 </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
 <td><code><a href="com/c77/androidstreamingclient/lib/rtp/RtpMediaJitterBuffer.html#FRAMES_WINDOW_PROPERTY">FRAMES_WINDOW_PROPERTY</a></code></td>
