+++ example/src/main/assets/configuration.ini
@@ -2,11 +2,9 @@
 DEBUGGING=true
 USE_NIO=true
 RECEIVE_BUFFER_SIZE_BYTES=50000
-# Has to be one of: 'fixed-frame-number', 'fixed-time' or 'zero-delay'
-BUFFER_TYPE=zero-delay
-# Window size for the fixed-frame-buffer:
-FRAMES_WINDOW_K=20
-# Maximum delay in milliseconds for the zero-delay buffer
-NODELAY_TIMEOUT=1000
-# Window size in milliseconds for the fixed-time buffer
+# Has to be one of: 'time-window' or 'min-delay'
+BUFFER_TYPE=time-window
+# Maximum delay in milliseconds for the min-delay buffer
+NODELAY_TIMEOUT=500
+# Window size in milliseconds for the time-window buffer
 FRAMES_WINDOW_TIME=1000
